pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "uncork"
version = "0.1.0"
description = "Uncork your Wine apps - convert Wine prefixes into portable Linux system packages"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["wine", "packaging", "linux", "deb", "pacman", "rpm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Software Distribution",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "icoextract>=0.1.4",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]
gui = [
    "PySide6>=6.5",
]

[project.scripts]
uncork = "uncork.cli:cli"

[project.urls]
Homepage = "https://github.com/yourname/uncork"
Repository = "https://github.com/yourname/uncork"
Issues = "https://github.com/yourname/uncork/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/uncork"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
